Refactor Firebase Admin initialization and update imports
- Updated Firebase Admin imports in login and register API routes to use new chunk.
- Modified dashboard API route to reflect the new Firebase Admin chunk.
- Refactored firebase-admin.ts to streamline initialization and improve error handling.
- Simplified environment variable validation and logging during Firebase Admin setup.
- Updated firebase.ts to clarify configuration comments and ensure analytics is only initialized in the browser.
