--- .vercel/output/_functions/manifest_BOaE-K_2.mjs
@@ -1,102 +1,0 @@
-import 'kleur/colors';
-import { g as decodeKey } from './chunks/astro/server_Bbe2uEvn.mjs';
-import 'clsx';
-import 'cookie';
-import { N as NOOP_MIDDLEWARE_FN } from './chunks/astro-designed-error-pages_C8u91leq.mjs';
-import 'es-module-lexer';
-
-function sanitizeParams(params) {
-  return Object.fromEntries(
-    Object.entries(params).map(([key, value]) => {
-      if (typeof value === "string") {
-        return [key, value.normalize().replace(/#/g, "%23").replace(/\?/g, "%3F")];
-      }
-      return [key, value];
-    })
-  );
-}
-function getParameter(part, params) {
-  if (part.spread) {
-    return params[part.content.slice(3)] || "";
-  }
-  if (part.dynamic) {
-    if (!params[part.content]) {
-      throw new TypeError(`Missing parameter: ${part.content}`);
-    }
-    return params[part.content];
-  }
-  return part.content.normalize().replace(/\?/g, "%3F").replace(/#/g, "%23").replace(/%5B/g, "[").replace(/%5D/g, "]");
-}
-function getSegment(segment, params) {
-  const segmentPath = segment.map((part) => getParameter(part, params)).join("");
-  return segmentPath ? "/" + segmentPath : "";
-}
-function getRouteGenerator(segments, addTrailingSlash) {
-  return (params) => {
-    const sanitizedParams = sanitizeParams(params);
-    let trailing = "";
-    if (addTrailingSlash === "always" && segments.length) {
-      trailing = "/";
-    }
-    const path = segments.map((segment) => getSegment(segment, sanitizedParams)).join("") + trailing;
-    return path || "/";
-  };
-}
-
-function deserializeRouteData(rawRouteData) {
-  return {
-    route: rawRouteData.route,
-    type: rawRouteData.type,
-    pattern: new RegExp(rawRouteData.pattern),
-    params: rawRouteData.params,
-    component: rawRouteData.component,
-    generate: getRouteGenerator(rawRouteData.segments, rawRouteData._meta.trailingSlash),
-    pathname: rawRouteData.pathname || void 0,
-    segments: rawRouteData.segments,
-    prerender: rawRouteData.prerender,
-    redirect: rawRouteData.redirect,
-    redirectRoute: rawRouteData.redirectRoute ? deserializeRouteData(rawRouteData.redirectRoute) : void 0,
-    fallbackRoutes: rawRouteData.fallbackRoutes.map((fallback) => {
-      return deserializeRouteData(fallback);
-    }),
-    isIndex: rawRouteData.isIndex,
-    origin: rawRouteData.origin
-  };
-}
-
-function deserializeManifest(serializedManifest) {
-  const routes = [];
-  for (const serializedRoute of serializedManifest.routes) {
-    routes.push({
-      ...serializedRoute,
-      routeData: deserializeRouteData(serializedRoute.routeData)
-    });
-    const route = serializedRoute;
-    route.routeData = deserializeRouteData(serializedRoute.routeData);
-  }
-  const assets = new Set(serializedManifest.assets);
-  const componentMetadata = new Map(serializedManifest.componentMetadata);
-  const inlinedScripts = new Map(serializedManifest.inlinedScripts);
-  const clientDirectives = new Map(serializedManifest.clientDirectives);
-  const serverIslandNameMap = new Map(serializedManifest.serverIslandNameMap);
-  const key = decodeKey(serializedManifest.key);
-  return {
-    // in case user middleware exists, this no-op middleware will be reassigned (see plugin-ssr.ts)
-    middleware() {
-      return { onRequest: NOOP_MIDDLEWARE_FN };
-    },
-    ...serializedManifest,
-    assets,
-    componentMetadata,
-    inlinedScripts,
-    clientDirectives,
-    routes,
-    serverIslandNameMap,
-    key
-  };
-}
-
-const manifest = deserializeManifest({"hrefRoot":"file:///D:/project/web/tukangcs/absent-accretion/","cacheDir":"file:///D:/project/web/tukangcs/absent-accretion/node_modules/.astro/","outDir":"file:///D:/project/web/tukangcs/absent-accretion/dist/","srcDir":"file:///D:/project/web/tukangcs/absent-accretion/src/","publicDir":"file:///D:/project/web/tukangcs/absent-accretion/public/","buildClientDir":"file:///D:/project/web/tukangcs/absent-accretion/dist/client/","buildServerDir":"file:///D:/project/web/tukangcs/absent-accretion/dist/server/","adapterName":"@astrojs/vercel","routes":[{"file":"","links":[],"scripts":[],"styles":[],"routeData":{"type":"page","component":"_server-islands.astro","params":["name"],"segments":[[{"content":"_server-islands","dynamic":false,"spread":false}],[{"content":"name","dynamic":true,"spread":false}]],"pattern":"^\\/_server-islands\\/([^/]+?)\\/?$","prerender":false,"isIndex":false,"fallbackRoutes":[],"route":"/_server-islands/[name]","origin":"internal","_meta":{"trailingSlash":"ignore"}}},{"file":"","links":[],"scripts":[{"stage":"head-inline","children":"window.va = window.va || function () { (window.vaq = window.vaq || []).push(arguments); };\n\t\tvar script = document.createElement('script');\n\t\tscript.defer = true;\n\t\tscript.src = '/_vercel/insights/script.js';\n\t\tvar head = document.querySelector('head');\n\t\thead.appendChild(script);\n\t"}],"styles":[],"routeData":{"type":"endpoint","isIndex":false,"route":"/_image","pattern":"^\\/_image\\/?$","segments":[[{"content":"_image","dynamic":false,"spread":false}]],"params":[],"component":"node_modules/astro/dist/assets/endpoint/generic.js","pathname":"/_image","prerender":false,"fallbackRoutes":[],"origin":"internal","_meta":{"trailingSlash":"ignore"}}},{"file":"","links":[],"scripts":[{"stage":"head-inline","children":"window.va = window.va || function () { (window.vaq = window.vaq || []).push(arguments); };\n\t\tvar script = document.createElement('script');\n\t\tscript.defer = true;\n\t\tscript.src = '/_vercel/insights/script.js';\n\t\tvar head = document.querySelector('head');\n\t\thead.appendChild(script);\n\t"}],"styles":[{"type":"external","src":"/_astro/about.yPWawPvy.css"},{"type":"inline","content":".footer-container[data-astro-cid-tnm6jnzf]{background:linear-gradient(135deg,#0a0a0a,#1a1a1a,#0f2027);color:#fff;position:relative;overflow:hidden}.footer-container[data-astro-cid-tnm6jnzf]:before{content:\"\";position:absolute;top:0;left:0;right:0;height:1px;background:linear-gradient(90deg,transparent,rgba(0,212,170,.3),transparent)}.footer-content[data-astro-cid-tnm6jnzf]{max-width:1200px;margin:0 auto;padding:4rem 2rem 2rem}.footer-grid[data-astro-cid-tnm6jnzf]{display:grid;grid-template-columns:2fr 1fr 1fr 1fr;gap:3rem;margin-bottom:3rem}.footer-section[data-astro-cid-tnm6jnzf]{animation:fadeInUp .6s ease-out}.footer-section[data-astro-cid-tnm6jnzf]:nth-child(2){animation-delay:.1s}.footer-section[data-astro-cid-tnm6jnzf]:nth-child(3){animation-delay:.2s}.footer-section[data-astro-cid-tnm6jnzf]:nth-child(4){animation-delay:.3s}.footer-logo[data-astro-cid-tnm6jnzf]{display:flex;align-items:center;gap:.75rem;margin-bottom:1.5rem}.logo-container[data-astro-cid-tnm6jnzf]{position:relative;width:48px;height:48px;flex-shrink:0}.logo-icon[data-astro-cid-tnm6jnzf]{width:100%;height:100%;filter:drop-shadow(0 0 8px rgba(99,102,241,.3))}.brand-content[data-astro-cid-tnm6jnzf]{display:flex;flex-direction:column}.brand-text[data-astro-cid-tnm6jnzf]{font-size:1.5rem;font-weight:700;color:#fff;letter-spacing:-.025em;line-height:1}.brand-tagline[data-astro-cid-tnm6jnzf]{font-size:.75rem;color:#fff9;margin-top:2px;font-weight:400}.footer-description[data-astro-cid-tnm6jnzf]{color:#ffffffb3;line-height:1.6;margin-bottom:1.5rem;font-size:.95rem}.social-links[data-astro-cid-tnm6jnzf]{display:flex;gap:1rem}.social-link[data-astro-cid-tnm6jnzf]{width:40px;height:40px;border-radius:50%;background:#ffffff0d;border:1px solid rgba(255,255,255,.1);display:flex;align-items:center;justify-content:center;color:#fff9;text-decoration:none;transition:all .3s ease;backdrop-filter:blur(10px)}.social-link[data-astro-cid-tnm6jnzf]:hover{background:#00d4aa1a;border-color:#00d4aa;color:#00d4aa;transform:translateY(-2px)}.footer-title[data-astro-cid-tnm6jnzf]{font-size:1rem;font-weight:600;margin-bottom:1rem;color:#ffffffe6}.footer-links[data-astro-cid-tnm6jnzf]{list-style:none;padding:0;margin:0}.footer-links[data-astro-cid-tnm6jnzf] li[data-astro-cid-tnm6jnzf]{margin-bottom:.75rem}.footer-links[data-astro-cid-tnm6jnzf] a[data-astro-cid-tnm6jnzf]{color:#fff9;text-decoration:none;font-size:.9rem;transition:color .3s ease}.footer-links[data-astro-cid-tnm6jnzf] a[data-astro-cid-tnm6jnzf]:hover{color:#00d4aa}.footer-bottom[data-astro-cid-tnm6jnzf]{border-top:1px solid rgba(255,255,255,.1);padding-top:2rem}.footer-bottom-content[data-astro-cid-tnm6jnzf]{display:flex;justify-content:space-between;align-items:center;flex-wrap:wrap;gap:1rem}.copyright[data-astro-cid-tnm6jnzf]{color:#ffffff80;font-size:.85rem;margin:0}.footer-legal[data-astro-cid-tnm6jnzf]{display:flex;gap:1.5rem}.footer-legal[data-astro-cid-tnm6jnzf] a[data-astro-cid-tnm6jnzf]{color:#ffffff80;text-decoration:none;font-size:.85rem;transition:color .3s ease}.footer-legal[data-astro-cid-tnm6jnzf] a[data-astro-cid-tnm6jnzf]:hover{color:#00d4aa}@media (max-width: 768px){.footer-content[data-astro-cid-tnm6jnzf]{padding:3rem 1rem 1rem}.footer-grid[data-astro-cid-tnm6jnzf]{grid-template-columns:1fr;gap:2rem;margin-bottom:2rem}.footer-section[data-astro-cid-tnm6jnzf]{text-align:center}.footer-logo[data-astro-cid-tnm6jnzf],.social-links[data-astro-cid-tnm6jnzf]{justify-content:center}.footer-bottom-content[data-astro-cid-tnm6jnzf]{flex-direction:column;text-align:center}.footer-legal[data-astro-cid-tnm6jnzf]{justify-content:center}}@media (max-width: 480px){.footer-legal[data-astro-cid-tnm6jnzf]{flex-direction:column;gap:.5rem}.logo-container[data-astro-cid-tnm6jnzf]{width:40px;height:40px}.brand-text[data-astro-cid-tnm6jnzf]{font-size:1.25rem}}@keyframes fadeInUp{0%{opacity:0;transform:translateY(20px)}to{opacity:1;transform:translateY(0)}}\n"},{"type":"external","src":"/_astro/about.CNm0yoeI.css"}],"routeData":{"route":"/about","isIndex":false,"type":"page","pattern":"^\\/about\\/?$","segments":[[{"content":"about","dynamic":false,"spread":false}]],"params":[],"component":"src/pages/about.astro","pathname":"/about","prerender":false,"fallbackRoutes":[],"distURL":[],"origin":"project","_meta":{"trailingSlash":"ignore"}}},{"file":"","links":[],"scripts":[{"stage":"head-inline","children":"window.va = window.va || function () { (window.vaq = window.vaq || []).push(arguments); };\n\t\tvar script = document.createElement('script');\n\t\tscript.defer = true;\n\t\tscript.src = '/_vercel/insights/script.js';\n\t\tvar head = document.querySelector('head');\n\t\thead.appendChild(script);\n\t"}],"styles":[],"routeData":{"route":"/api/auth/login","isIndex":false,"type":"endpoint","pattern":"^\\/api\\/auth\\/login\\/?$","segments":[[{"content":"api","dynamic":false,"spread":false}],[{"content":"auth","dynamic":false,"spread":false}],[{"content":"login","dynamic":false,"spread":false}]],"params":[],"component":"src/pages/api/auth/login.ts","pathname":"/api/auth/login","prerender":false,"fallbackRoutes":[],"distURL":[],"origin":"project","_meta":{"trailingSlash":"ignore"}}},{"file":"","links":[],"scripts":[{"stage":"head-inline","children":"window.va = window.va || function () { (window.vaq = window.vaq || []).push(arguments); };\n\t\tvar script = document.createElement('script');\n\t\tscript.defer = true;\n\t\tscript.src = '/_vercel/insights/script.js';\n\t\tvar head = document.querySelector('head');\n\t\thead.appendChild(script);\n\t"}],"styles":[],"routeData":{"route":"/api/auth/register","isIndex":false,"type":"endpoint","pattern":"^\\/api\\/auth\\/register\\/?$","segments":[[{"content":"api","dynamic":false,"spread":false}],[{"content":"auth","dynamic":false,"spread":false}],[{"content":"register","dynamic":false,"spread":false}]],"params":[],"component":"src/pages/api/auth/register.ts","pathname":"/api/auth/register","prerender":false,"fallbackRoutes":[],"distURL":[],"origin":"project","_meta":{"trailingSlash":"ignore"}}},{"file":"","links":[],"scripts":[{"stage":"head-inline","children":"window.va = window.va || function () { (window.vaq = window.vaq || []).push(arguments); };\n\t\tvar script = document.createElement('script');\n\t\tscript.defer = true;\n\t\tscript.src = '/_vercel/insights/script.js';\n\t\tvar head = document.querySelector('head');\n\t\thead.appendChild(script);\n\t"}],"styles":[],"routeData":{"route":"/api/auth/signout","isIndex":false,"type":"endpoint","pattern":"^\\/api\\/auth\\/signout\\/?$","segments":[[{"content":"api","dynamic":false,"spread":false}],[{"content":"auth","dynamic":false,"spread":false}],[{"content":"signout","dynamic":false,"spread":false}]],"params":[],"component":"src/pages/api/auth/signout.ts","pathname":"/api/auth/signout","prerender":false,"fallbackRoutes":[],"distURL":[],"origin":"project","_meta":{"trailingSlash":"ignore"}}},{"file":"","links":[],"scripts":[{"stage":"head-inline","children":"window.va = window.va || function () { (window.vaq = window.vaq || []).push(arguments); };\n\t\tvar script = document.createElement('script');\n\t\tscript.defer = true;\n\t\tscript.src = '/_vercel/insights/script.js';\n\t\tvar head = document.querySelector('head');\n\t\thead.appendChild(script);\n\t"}],"styles":[],"routeData":{"route":"/api/test-env","isIndex":false,"type":"endpoint","pattern":"^\\/api\\/test-env\\/?$","segments":[[{"content":"api","dynamic":false,"spread":false}],[{"content":"test-env","dynamic":false,"spread":false}]],"params":[],"component":"src/pages/api/test-env.ts","pathname":"/api/test-env","prerender":false,"fallbackRoutes":[],"distURL":[],"origin":"project","_meta":{"trailingSlash":"ignore"}}},{"file":"","links":[],"scripts":[{"stage":"head-inline","children":"window.va = window.va || function () { (window.vaq = window.vaq || []).push(arguments); };\n\t\tvar script = document.createElement('script');\n\t\tscript.defer = true;\n\t\tscript.src = '/_vercel/insights/script.js';\n\t\tvar head = document.querySelector('head');\n\t\thead.appendChild(script);\n\t"}],"styles":[{"type":"inline","content":".footer-container[data-astro-cid-tnm6jnzf]{background:linear-gradient(135deg,#0a0a0a,#1a1a1a,#0f2027);color:#fff;position:relative;overflow:hidden}.footer-container[data-astro-cid-tnm6jnzf]:before{content:\"\";position:absolute;top:0;left:0;right:0;height:1px;background:linear-gradient(90deg,transparent,rgba(0,212,170,.3),transparent)}.footer-content[data-astro-cid-tnm6jnzf]{max-width:1200px;margin:0 auto;padding:4rem 2rem 2rem}.footer-grid[data-astro-cid-tnm6jnzf]{display:grid;grid-template-columns:2fr 1fr 1fr 1fr;gap:3rem;margin-bottom:3rem}.footer-section[data-astro-cid-tnm6jnzf]{animation:fadeInUp .6s ease-out}.footer-section[data-astro-cid-tnm6jnzf]:nth-child(2){animation-delay:.1s}.footer-section[data-astro-cid-tnm6jnzf]:nth-child(3){animation-delay:.2s}.footer-section[data-astro-cid-tnm6jnzf]:nth-child(4){animation-delay:.3s}.footer-logo[data-astro-cid-tnm6jnzf]{display:flex;align-items:center;gap:.75rem;margin-bottom:1.5rem}.logo-container[data-astro-cid-tnm6jnzf]{position:relative;width:48px;height:48px;flex-shrink:0}.logo-icon[data-astro-cid-tnm6jnzf]{width:100%;height:100%;filter:drop-shadow(0 0 8px rgba(99,102,241,.3))}.brand-content[data-astro-cid-tnm6jnzf]{display:flex;flex-direction:column}.brand-text[data-astro-cid-tnm6jnzf]{font-size:1.5rem;font-weight:700;color:#fff;letter-spacing:-.025em;line-height:1}.brand-tagline[data-astro-cid-tnm6jnzf]{font-size:.75rem;color:#fff9;margin-top:2px;font-weight:400}.footer-description[data-astro-cid-tnm6jnzf]{color:#ffffffb3;line-height:1.6;margin-bottom:1.5rem;font-size:.95rem}.social-links[data-astro-cid-tnm6jnzf]{display:flex;gap:1rem}.social-link[data-astro-cid-tnm6jnzf]{width:40px;height:40px;border-radius:50%;background:#ffffff0d;border:1px solid rgba(255,255,255,.1);display:flex;align-items:center;justify-content:center;color:#fff9;text-decoration:none;transition:all .3s ease;backdrop-filter:blur(10px)}.social-link[data-astro-cid-tnm6jnzf]:hover{background:#00d4aa1a;border-color:#00d4aa;color:#00d4aa;transform:translateY(-2px)}.footer-title[data-astro-cid-tnm6jnzf]{font-size:1rem;font-weight:600;margin-bottom:1rem;color:#ffffffe6}.footer-links[data-astro-cid-tnm6jnzf]{list-style:none;padding:0;margin:0}.footer-links[data-astro-cid-tnm6jnzf] li[data-astro-cid-tnm6jnzf]{margin-bottom:.75rem}.footer-links[data-astro-cid-tnm6jnzf] a[data-astro-cid-tnm6jnzf]{color:#fff9;text-decoration:none;font-size:.9rem;transition:color .3s ease}.footer-links[data-astro-cid-tnm6jnzf] a[data-astro-cid-tnm6jnzf]:hover{color:#00d4aa}.footer-bottom[data-astro-cid-tnm6jnzf]{border-top:1px solid rgba(255,255,255,.1);padding-top:2rem}.footer-bottom-content[data-astro-cid-tnm6jnzf]{display:flex;justify-content:space-between;align-items:center;flex-wrap:wrap;gap:1rem}.copyright[data-astro-cid-tnm6jnzf]{color:#ffffff80;font-size:.85rem;margin:0}.footer-legal[data-astro-cid-tnm6jnzf]{display:flex;gap:1.5rem}.footer-legal[data-astro-cid-tnm6jnzf] a[data-astro-cid-tnm6jnzf]{color:#ffffff80;text-decoration:none;font-size:.85rem;transition:color .3s ease}.footer-legal[data-astro-cid-tnm6jnzf] a[data-astro-cid-tnm6jnzf]:hover{color:#00d4aa}@media (max-width: 768px){.footer-content[data-astro-cid-tnm6jnzf]{padding:3rem 1rem 1rem}.footer-grid[data-astro-cid-tnm6jnzf]{grid-template-columns:1fr;gap:2rem;margin-bottom:2rem}.footer-section[data-astro-cid-tnm6jnzf]{text-align:center}.footer-logo[data-astro-cid-tnm6jnzf],.social-links[data-astro-cid-tnm6jnzf]{justify-content:center}.footer-bottom-content[data-astro-cid-tnm6jnzf]{flex-direction:column;text-align:center}.footer-legal[data-astro-cid-tnm6jnzf]{justify-content:center}}@media (max-width: 480px){.footer-legal[data-astro-cid-tnm6jnzf]{flex-direction:column;gap:.5rem}.logo-container[data-astro-cid-tnm6jnzf]{width:40px;height:40px}.brand-text[data-astro-cid-tnm6jnzf]{font-size:1.25rem}}@keyframes fadeInUp{0%{opacity:0;transform:translateY(20px)}to{opacity:1;transform:translateY(0)}}\n"},{"type":"external","src":"/_astro/about.yPWawPvy.css"},{"type":"external","src":"/_astro/contact.BeCkmT8s.css"}],"routeData":{"route":"/contact","isIndex":false,"type":"page","pattern":"^\\/contact\\/?$","segments":[[{"content":"contact","dynamic":false,"spread":false}]],"params":[],"component":"src/pages/contact.astro","pathname":"/contact","prerender":false,"fallbackRoutes":[],"distURL":[],"origin":"project","_meta":{"trailingSlash":"ignore"}}},{"file":"","links":[],"scripts":[{"stage":"head-inline","children":"window.va = window.va || function () { (window.vaq = window.vaq || []).push(arguments); };\n\t\tvar script = document.createElement('script');\n\t\tscript.defer = true;\n\t\tscript.src = '/_vercel/insights/script.js';\n\t\tvar head = document.querySelector('head');\n\t\thead.appendChild(script);\n\t"}],"styles":[{"type":"inline","content":"[data-astro-cid-3nssi2tu]{margin:0;padding:0;box-sizing:border-box}body{font-family:Segoe UI,Tahoma,Geneva,Verdana,sans-serif;background:linear-gradient(135deg,#667eea,#764ba2);min-height:100vh;color:#333}.header[data-astro-cid-3nssi2tu]{background:#ffffff1a;backdrop-filter:blur(10px);padding:1rem 2rem;display:flex;justify-content:space-between;align-items:center;border-bottom:1px solid rgba(255,255,255,.2)}.logo[data-astro-cid-3nssi2tu]{font-size:1.5rem;font-weight:700;color:#fff}.logout-btn[data-astro-cid-3nssi2tu]{background:#fff3;color:#fff;border:none;padding:.5rem 1rem;border-radius:8px;cursor:pointer;transition:all .3s ease}.logout-btn[data-astro-cid-3nssi2tu]:hover{background:#ffffff4d}.container[data-astro-cid-3nssi2tu]{max-width:1200px;margin:2rem auto;padding:0 2rem}.welcome-card[data-astro-cid-3nssi2tu]{background:#fff;border-radius:12px;padding:2rem;box-shadow:0 8px 32px #0000001a;margin-bottom:2rem}.welcome-title[data-astro-cid-3nssi2tu]{font-size:2rem;margin-bottom:.5rem;color:#333}.welcome-subtitle[data-astro-cid-3nssi2tu]{color:#666;margin-bottom:1rem}.user-info[data-astro-cid-3nssi2tu]{background:#f8f9fa;padding:1rem;border-radius:8px;margin-top:1rem}.user-info[data-astro-cid-3nssi2tu] h3[data-astro-cid-3nssi2tu]{margin-bottom:.5rem;color:#333}.user-info[data-astro-cid-3nssi2tu] p[data-astro-cid-3nssi2tu]{margin-bottom:.25rem;color:#666}.features-grid[data-astro-cid-3nssi2tu]{display:grid;grid-template-columns:repeat(auto-fit,minmax(300px,1fr));gap:2rem;margin-top:2rem}.feature-card[data-astro-cid-3nssi2tu]{background:#fff;border-radius:12px;padding:2rem;box-shadow:0 8px 32px #0000001a;text-align:center;transition:transform .3s ease}.feature-card[data-astro-cid-3nssi2tu]:hover{transform:translateY(-5px)}.feature-icon[data-astro-cid-3nssi2tu]{font-size:3rem;margin-bottom:1rem}.feature-title[data-astro-cid-3nssi2tu]{font-size:1.25rem;font-weight:700;margin-bottom:.5rem;color:#333}.feature-description[data-astro-cid-3nssi2tu]{color:#666;line-height:1.6}.coming-soon[data-astro-cid-3nssi2tu]{background:linear-gradient(45deg,#ff6b6b,orange);color:#fff;padding:.25rem .5rem;border-radius:4px;font-size:.75rem;margin-left:.5rem}\n"}],"routeData":{"route":"/dashboard","isIndex":false,"type":"page","pattern":"^\\/dashboard\\/?$","segments":[[{"content":"dashboard","dynamic":false,"spread":false}]],"params":[],"component":"src/pages/dashboard.astro","pathname":"/dashboard","prerender":false,"fallbackRoutes":[],"distURL":[],"origin":"project","_meta":{"trailingSlash":"ignore"}}},{"file":"","links":[],"scripts":[{"stage":"head-inline","children":"window.va = window.va || function () { (window.vaq = window.vaq || []).push(arguments); };\n\t\tvar script = document.createElement('script');\n\t\tscript.defer = true;\n\t\tscript.src = '/_vercel/insights/script.js';\n\t\tvar head = document.querySelector('head');\n\t\thead.appendChild(script);\n\t"}],"styles":[{"type":"external","src":"/_astro/about.yPWawPvy.css"},{"type":"inline","content":".footer-container[data-astro-cid-tnm6jnzf]{background:linear-gradient(135deg,#0a0a0a,#1a1a1a,#0f2027);color:#fff;position:relative;overflow:hidden}.footer-container[data-astro-cid-tnm6jnzf]:before{content:\"\";position:absolute;top:0;left:0;right:0;height:1px;background:linear-gradient(90deg,transparent,rgba(0,212,170,.3),transparent)}.footer-content[data-astro-cid-tnm6jnzf]{max-width:1200px;margin:0 auto;padding:4rem 2rem 2rem}.footer-grid[data-astro-cid-tnm6jnzf]{display:grid;grid-template-columns:2fr 1fr 1fr 1fr;gap:3rem;margin-bottom:3rem}.footer-section[data-astro-cid-tnm6jnzf]{animation:fadeInUp .6s ease-out}.footer-section[data-astro-cid-tnm6jnzf]:nth-child(2){animation-delay:.1s}.footer-section[data-astro-cid-tnm6jnzf]:nth-child(3){animation-delay:.2s}.footer-section[data-astro-cid-tnm6jnzf]:nth-child(4){animation-delay:.3s}.footer-logo[data-astro-cid-tnm6jnzf]{display:flex;align-items:center;gap:.75rem;margin-bottom:1.5rem}.logo-container[data-astro-cid-tnm6jnzf]{position:relative;width:48px;height:48px;flex-shrink:0}.logo-icon[data-astro-cid-tnm6jnzf]{width:100%;height:100%;filter:drop-shadow(0 0 8px rgba(99,102,241,.3))}.brand-content[data-astro-cid-tnm6jnzf]{display:flex;flex-direction:column}.brand-text[data-astro-cid-tnm6jnzf]{font-size:1.5rem;font-weight:700;color:#fff;letter-spacing:-.025em;line-height:1}.brand-tagline[data-astro-cid-tnm6jnzf]{font-size:.75rem;color:#fff9;margin-top:2px;font-weight:400}.footer-description[data-astro-cid-tnm6jnzf]{color:#ffffffb3;line-height:1.6;margin-bottom:1.5rem;font-size:.95rem}.social-links[data-astro-cid-tnm6jnzf]{display:flex;gap:1rem}.social-link[data-astro-cid-tnm6jnzf]{width:40px;height:40px;border-radius:50%;background:#ffffff0d;border:1px solid rgba(255,255,255,.1);display:flex;align-items:center;justify-content:center;color:#fff9;text-decoration:none;transition:all .3s ease;backdrop-filter:blur(10px)}.social-link[data-astro-cid-tnm6jnzf]:hover{background:#00d4aa1a;border-color:#00d4aa;color:#00d4aa;transform:translateY(-2px)}.footer-title[data-astro-cid-tnm6jnzf]{font-size:1rem;font-weight:600;margin-bottom:1rem;color:#ffffffe6}.footer-links[data-astro-cid-tnm6jnzf]{list-style:none;padding:0;margin:0}.footer-links[data-astro-cid-tnm6jnzf] li[data-astro-cid-tnm6jnzf]{margin-bottom:.75rem}.footer-links[data-astro-cid-tnm6jnzf] a[data-astro-cid-tnm6jnzf]{color:#fff9;text-decoration:none;font-size:.9rem;transition:color .3s ease}.footer-links[data-astro-cid-tnm6jnzf] a[data-astro-cid-tnm6jnzf]:hover{color:#00d4aa}.footer-bottom[data-astro-cid-tnm6jnzf]{border-top:1px solid rgba(255,255,255,.1);padding-top:2rem}.footer-bottom-content[data-astro-cid-tnm6jnzf]{display:flex;justify-content:space-between;align-items:center;flex-wrap:wrap;gap:1rem}.copyright[data-astro-cid-tnm6jnzf]{color:#ffffff80;font-size:.85rem;margin:0}.footer-legal[data-astro-cid-tnm6jnzf]{display:flex;gap:1.5rem}.footer-legal[data-astro-cid-tnm6jnzf] a[data-astro-cid-tnm6jnzf]{color:#ffffff80;text-decoration:none;font-size:.85rem;transition:color .3s ease}.footer-legal[data-astro-cid-tnm6jnzf] a[data-astro-cid-tnm6jnzf]:hover{color:#00d4aa}@media (max-width: 768px){.footer-content[data-astro-cid-tnm6jnzf]{padding:3rem 1rem 1rem}.footer-grid[data-astro-cid-tnm6jnzf]{grid-template-columns:1fr;gap:2rem;margin-bottom:2rem}.footer-section[data-astro-cid-tnm6jnzf]{text-align:center}.footer-logo[data-astro-cid-tnm6jnzf],.social-links[data-astro-cid-tnm6jnzf]{justify-content:center}.footer-bottom-content[data-astro-cid-tnm6jnzf]{flex-direction:column;text-align:center}.footer-legal[data-astro-cid-tnm6jnzf]{justify-content:center}}@media (max-width: 480px){.footer-legal[data-astro-cid-tnm6jnzf]{flex-direction:column;gap:.5rem}.logo-container[data-astro-cid-tnm6jnzf]{width:40px;height:40px}.brand-text[data-astro-cid-tnm6jnzf]{font-size:1.25rem}}@keyframes fadeInUp{0%{opacity:0;transform:translateY(20px)}to{opacity:1;transform:translateY(0)}}\nbody{margin:0;font-family:Segoe UI,Tahoma,Geneva,Verdana,sans-serif;background:linear-gradient(to bottom right,#111827,#581c87,#1e3a8a);color:#fff;min-height:100vh;overflow-x:hidden}.container{max-width:600px;margin:6rem auto;padding:2rem;background:#ffffff0d;backdrop-filter:blur(16px);border:1px solid rgba(255,255,255,.1);border-radius:1.5rem;box-shadow:0 0 30px #6366f133}h1{font-size:2rem;text-align:center;margin-bottom:1rem;background:linear-gradient(135deg,#8b5cf6,#6366f1);-webkit-background-clip:text;-webkit-text-fill-color:transparent}.tabs{display:flex;justify-content:center;margin-bottom:2rem}.tab{padding:.75rem 1.5rem;cursor:pointer;background:transparent;border:none;font-weight:700;color:#cbd5e1;border-bottom:2px solid transparent;transition:all .3s ease}.tab.active{color:#fff;border-color:#8b5cf6}form{display:flex;flex-direction:column;gap:1rem}input,select{padding:.75rem 1rem;border-radius:.75rem;border:none;background:#ffffff1a;color:#fff;font-size:1rem;border:1px solid rgba(255,255,255,.2)}input::placeholder{color:#ffffff80}input:focus,select:focus{outline:none;border-color:#8b5cf6;box-shadow:0 0 0 2px #8b5cf680}button{padding:.75rem 1rem;font-weight:700;border:none;border-radius:.75rem;background:linear-gradient(135deg,#6366f1,#8b5cf6);color:#fff;cursor:pointer;transition:transform .3s ease,box-shadow .3s ease}button:hover:not(:disabled){transform:translateY(-2px);box-shadow:0 8px 24px #6366f166}button:disabled{opacity:.6;cursor:not-allowed;transform:none}.note{text-align:center;font-size:.9rem;color:#94a3b8}.hidden{display:none}.message{padding:.75rem;border-radius:.75rem;margin-bottom:1rem;text-align:center;font-weight:700}.success{background:#22c55e33;border:1px solid rgba(34,197,94,.3);color:#4ade80}.error{background:#ef444433;border:1px solid rgba(239,68,68,.3);color:#f87171}select[multiple]{height:auto;min-height:100px}select option{background:#1e293b;color:#fff;padding:.5rem}label{color:#cbd5e1;font-size:.9rem;margin-bottom:.5rem}.loading{display:inline-block;width:16px;height:16px;border:2px solid rgba(255,255,255,.3);border-top:2px solid white;border-radius:50%;animation:spin 1s linear infinite;margin-right:8px}@keyframes spin{0%{transform:rotate(0)}to{transform:rotate(360deg)}}\n"}],"routeData":{"route":"/get-started","isIndex":false,"type":"page","pattern":"^\\/get-started\\/?$","segments":[[{"content":"get-started","dynamic":false,"spread":false}]],"params":[],"component":"src/pages/get-started.astro","pathname":"/get-started","prerender":false,"fallbackRoutes":[],"distURL":[],"origin":"project","_meta":{"trailingSlash":"ignore"}}},{"file":"","links":[],"scripts":[{"stage":"head-inline","children":"window.va = window.va || function () { (window.vaq = window.vaq || []).push(arguments); };\n\t\tvar script = document.createElement('script');\n\t\tscript.defer = true;\n\t\tscript.src = '/_vercel/insights/script.js';\n\t\tvar head = document.querySelector('head');\n\t\thead.appendChild(script);\n\t"}],"styles":[{"type":"external","src":"/_astro/about.yPWawPvy.css"},{"type":"inline","content":".footer-container[data-astro-cid-tnm6jnzf]{background:linear-gradient(135deg,#0a0a0a,#1a1a1a,#0f2027);color:#fff;position:relative;overflow:hidden}.footer-container[data-astro-cid-tnm6jnzf]:before{content:\"\";position:absolute;top:0;left:0;right:0;height:1px;background:linear-gradient(90deg,transparent,rgba(0,212,170,.3),transparent)}.footer-content[data-astro-cid-tnm6jnzf]{max-width:1200px;margin:0 auto;padding:4rem 2rem 2rem}.footer-grid[data-astro-cid-tnm6jnzf]{display:grid;grid-template-columns:2fr 1fr 1fr 1fr;gap:3rem;margin-bottom:3rem}.footer-section[data-astro-cid-tnm6jnzf]{animation:fadeInUp .6s ease-out}.footer-section[data-astro-cid-tnm6jnzf]:nth-child(2){animation-delay:.1s}.footer-section[data-astro-cid-tnm6jnzf]:nth-child(3){animation-delay:.2s}.footer-section[data-astro-cid-tnm6jnzf]:nth-child(4){animation-delay:.3s}.footer-logo[data-astro-cid-tnm6jnzf]{display:flex;align-items:center;gap:.75rem;margin-bottom:1.5rem}.logo-container[data-astro-cid-tnm6jnzf]{position:relative;width:48px;height:48px;flex-shrink:0}.logo-icon[data-astro-cid-tnm6jnzf]{width:100%;height:100%;filter:drop-shadow(0 0 8px rgba(99,102,241,.3))}.brand-content[data-astro-cid-tnm6jnzf]{display:flex;flex-direction:column}.brand-text[data-astro-cid-tnm6jnzf]{font-size:1.5rem;font-weight:700;color:#fff;letter-spacing:-.025em;line-height:1}.brand-tagline[data-astro-cid-tnm6jnzf]{font-size:.75rem;color:#fff9;margin-top:2px;font-weight:400}.footer-description[data-astro-cid-tnm6jnzf]{color:#ffffffb3;line-height:1.6;margin-bottom:1.5rem;font-size:.95rem}.social-links[data-astro-cid-tnm6jnzf]{display:flex;gap:1rem}.social-link[data-astro-cid-tnm6jnzf]{width:40px;height:40px;border-radius:50%;background:#ffffff0d;border:1px solid rgba(255,255,255,.1);display:flex;align-items:center;justify-content:center;color:#fff9;text-decoration:none;transition:all .3s ease;backdrop-filter:blur(10px)}.social-link[data-astro-cid-tnm6jnzf]:hover{background:#00d4aa1a;border-color:#00d4aa;color:#00d4aa;transform:translateY(-2px)}.footer-title[data-astro-cid-tnm6jnzf]{font-size:1rem;font-weight:600;margin-bottom:1rem;color:#ffffffe6}.footer-links[data-astro-cid-tnm6jnzf]{list-style:none;padding:0;margin:0}.footer-links[data-astro-cid-tnm6jnzf] li[data-astro-cid-tnm6jnzf]{margin-bottom:.75rem}.footer-links[data-astro-cid-tnm6jnzf] a[data-astro-cid-tnm6jnzf]{color:#fff9;text-decoration:none;font-size:.9rem;transition:color .3s ease}.footer-links[data-astro-cid-tnm6jnzf] a[data-astro-cid-tnm6jnzf]:hover{color:#00d4aa}.footer-bottom[data-astro-cid-tnm6jnzf]{border-top:1px solid rgba(255,255,255,.1);padding-top:2rem}.footer-bottom-content[data-astro-cid-tnm6jnzf]{display:flex;justify-content:space-between;align-items:center;flex-wrap:wrap;gap:1rem}.copyright[data-astro-cid-tnm6jnzf]{color:#ffffff80;font-size:.85rem;margin:0}.footer-legal[data-astro-cid-tnm6jnzf]{display:flex;gap:1.5rem}.footer-legal[data-astro-cid-tnm6jnzf] a[data-astro-cid-tnm6jnzf]{color:#ffffff80;text-decoration:none;font-size:.85rem;transition:color .3s ease}.footer-legal[data-astro-cid-tnm6jnzf] a[data-astro-cid-tnm6jnzf]:hover{color:#00d4aa}@media (max-width: 768px){.footer-content[data-astro-cid-tnm6jnzf]{padding:3rem 1rem 1rem}.footer-grid[data-astro-cid-tnm6jnzf]{grid-template-columns:1fr;gap:2rem;margin-bottom:2rem}.footer-section[data-astro-cid-tnm6jnzf]{text-align:center}.footer-logo[data-astro-cid-tnm6jnzf],.social-links[data-astro-cid-tnm6jnzf]{justify-content:center}.footer-bottom-content[data-astro-cid-tnm6jnzf]{flex-direction:column;text-align:center}.footer-legal[data-astro-cid-tnm6jnzf]{justify-content:center}}@media (max-width: 480px){.footer-legal[data-astro-cid-tnm6jnzf]{flex-direction:column;gap:.5rem}.logo-container[data-astro-cid-tnm6jnzf]{width:40px;height:40px}.brand-text[data-astro-cid-tnm6jnzf]{font-size:1.25rem}}@keyframes fadeInUp{0%{opacity:0;transform:translateY(20px)}to{opacity:1;transform:translateY(0)}}\nbody{margin:0;font-family:Segoe UI,Tahoma,Geneva,Verdana,sans-serif;background:linear-gradient(to bottom right,#111827,#581c87,#1e3a8a);color:#fff;min-height:100vh;overflow-x:hidden}.container{max-width:600px;margin:6rem auto;padding:2rem;background:#ffffff0d;backdrop-filter:blur(16px);border:1px solid rgba(255,255,255,.1);border-radius:1.5rem;box-shadow:0 0 30px #6366f133}h1{font-size:2rem;text-align:center;margin-bottom:1.5rem;background:linear-gradient(135deg,#8b5cf6,#6366f1);-webkit-background-clip:text;-webkit-text-fill-color:transparent}form{display:flex;flex-direction:column;gap:1rem}input{padding:.75rem 1rem;border-radius:.75rem;border:none;background:#ffffff1a;color:#fff;font-size:1rem;border:1px solid rgba(255,255,255,.2)}input::placeholder{color:#ffffff80}input:focus{outline:none;border-color:#8b5cf6;box-shadow:0 0 0 2px #8b5cf680}button{padding:.75rem 1rem;font-weight:700;border:none;border-radius:.75rem;background:linear-gradient(135deg,#6366f1,#8b5cf6);color:#fff;cursor:pointer;transition:transform .3s ease,box-shadow .3s ease}button:hover{transform:translateY(-2px);box-shadow:0 8px 24px #6366f166}button:disabled{opacity:.6;cursor:not-allowed;transform:none}.note{text-align:center;font-size:.9rem;color:#94a3b8;margin-top:1rem}.note a{color:#8b5cf6;text-decoration:underline}.message{padding:.75rem;border-radius:.75rem;margin-bottom:1rem;text-align:center;font-weight:700}.success{background:#22c55e33;border:1px solid rgba(34,197,94,.3);color:#4ade80}.error{background:#ef444433;border:1px solid rgba(239,68,68,.3);color:#f87171}\n"}],"routeData":{"route":"/login","isIndex":false,"type":"page","pattern":"^\\/login\\/?$","segments":[[{"content":"login","dynamic":false,"spread":false}]],"params":[],"component":"src/pages/login.astro","pathname":"/login","prerender":false,"fallbackRoutes":[],"distURL":[],"origin":"project","_meta":{"trailingSlash":"ignore"}}},{"file":"","links":[],"scripts":[{"stage":"head-inline","children":"window.va = window.va || function () { (window.vaq = window.vaq || []).push(arguments); };\n\t\tvar script = document.createElement('script');\n\t\tscript.defer = true;\n\t\tscript.src = '/_vercel/insights/script.js';\n\t\tvar head = document.querySelector('head');\n\t\thead.appendChild(script);\n\t"}],"styles":[{"type":"external","src":"/_astro/about.yPWawPvy.css"},{"type":"inline","content":".footer-container[data-astro-cid-tnm6jnzf]{background:linear-gradient(135deg,#0a0a0a,#1a1a1a,#0f2027);color:#fff;position:relative;overflow:hidden}.footer-container[data-astro-cid-tnm6jnzf]:before{content:\"\";position:absolute;top:0;left:0;right:0;height:1px;background:linear-gradient(90deg,transparent,rgba(0,212,170,.3),transparent)}.footer-content[data-astro-cid-tnm6jnzf]{max-width:1200px;margin:0 auto;padding:4rem 2rem 2rem}.footer-grid[data-astro-cid-tnm6jnzf]{display:grid;grid-template-columns:2fr 1fr 1fr 1fr;gap:3rem;margin-bottom:3rem}.footer-section[data-astro-cid-tnm6jnzf]{animation:fadeInUp .6s ease-out}.footer-section[data-astro-cid-tnm6jnzf]:nth-child(2){animation-delay:.1s}.footer-section[data-astro-cid-tnm6jnzf]:nth-child(3){animation-delay:.2s}.footer-section[data-astro-cid-tnm6jnzf]:nth-child(4){animation-delay:.3s}.footer-logo[data-astro-cid-tnm6jnzf]{display:flex;align-items:center;gap:.75rem;margin-bottom:1.5rem}.logo-container[data-astro-cid-tnm6jnzf]{position:relative;width:48px;height:48px;flex-shrink:0}.logo-icon[data-astro-cid-tnm6jnzf]{width:100%;height:100%;filter:drop-shadow(0 0 8px rgba(99,102,241,.3))}.brand-content[data-astro-cid-tnm6jnzf]{display:flex;flex-direction:column}.brand-text[data-astro-cid-tnm6jnzf]{font-size:1.5rem;font-weight:700;color:#fff;letter-spacing:-.025em;line-height:1}.brand-tagline[data-astro-cid-tnm6jnzf]{font-size:.75rem;color:#fff9;margin-top:2px;font-weight:400}.footer-description[data-astro-cid-tnm6jnzf]{color:#ffffffb3;line-height:1.6;margin-bottom:1.5rem;font-size:.95rem}.social-links[data-astro-cid-tnm6jnzf]{display:flex;gap:1rem}.social-link[data-astro-cid-tnm6jnzf]{width:40px;height:40px;border-radius:50%;background:#ffffff0d;border:1px solid rgba(255,255,255,.1);display:flex;align-items:center;justify-content:center;color:#fff9;text-decoration:none;transition:all .3s ease;backdrop-filter:blur(10px)}.social-link[data-astro-cid-tnm6jnzf]:hover{background:#00d4aa1a;border-color:#00d4aa;color:#00d4aa;transform:translateY(-2px)}.footer-title[data-astro-cid-tnm6jnzf]{font-size:1rem;font-weight:600;margin-bottom:1rem;color:#ffffffe6}.footer-links[data-astro-cid-tnm6jnzf]{list-style:none;padding:0;margin:0}.footer-links[data-astro-cid-tnm6jnzf] li[data-astro-cid-tnm6jnzf]{margin-bottom:.75rem}.footer-links[data-astro-cid-tnm6jnzf] a[data-astro-cid-tnm6jnzf]{color:#fff9;text-decoration:none;font-size:.9rem;transition:color .3s ease}.footer-links[data-astro-cid-tnm6jnzf] a[data-astro-cid-tnm6jnzf]:hover{color:#00d4aa}.footer-bottom[data-astro-cid-tnm6jnzf]{border-top:1px solid rgba(255,255,255,.1);padding-top:2rem}.footer-bottom-content[data-astro-cid-tnm6jnzf]{display:flex;justify-content:space-between;align-items:center;flex-wrap:wrap;gap:1rem}.copyright[data-astro-cid-tnm6jnzf]{color:#ffffff80;font-size:.85rem;margin:0}.footer-legal[data-astro-cid-tnm6jnzf]{display:flex;gap:1.5rem}.footer-legal[data-astro-cid-tnm6jnzf] a[data-astro-cid-tnm6jnzf]{color:#ffffff80;text-decoration:none;font-size:.85rem;transition:color .3s ease}.footer-legal[data-astro-cid-tnm6jnzf] a[data-astro-cid-tnm6jnzf]:hover{color:#00d4aa}@media (max-width: 768px){.footer-content[data-astro-cid-tnm6jnzf]{padding:3rem 1rem 1rem}.footer-grid[data-astro-cid-tnm6jnzf]{grid-template-columns:1fr;gap:2rem;margin-bottom:2rem}.footer-section[data-astro-cid-tnm6jnzf]{text-align:center}.footer-logo[data-astro-cid-tnm6jnzf],.social-links[data-astro-cid-tnm6jnzf]{justify-content:center}.footer-bottom-content[data-astro-cid-tnm6jnzf]{flex-direction:column;text-align:center}.footer-legal[data-astro-cid-tnm6jnzf]{justify-content:center}}@media (max-width: 480px){.footer-legal[data-astro-cid-tnm6jnzf]{flex-direction:column;gap:.5rem}.logo-container[data-astro-cid-tnm6jnzf]{width:40px;height:40px}.brand-text[data-astro-cid-tnm6jnzf]{font-size:1.25rem}}@keyframes fadeInUp{0%{opacity:0;transform:translateY(20px)}to{opacity:1;transform:translateY(0)}}\n"},{"type":"external","src":"/_astro/product.DkOdxu3m.css"}],"routeData":{"route":"/product","isIndex":false,"type":"page","pattern":"^\\/product\\/?$","segments":[[{"content":"product","dynamic":false,"spread":false}]],"params":[],"component":"src/pages/product.astro","pathname":"/product","prerender":false,"fallbackRoutes":[],"distURL":[],"origin":"project","_meta":{"trailingSlash":"ignore"}}},{"file":"","links":[],"scripts":[{"stage":"head-inline","children":"window.va = window.va || function () { (window.vaq = window.vaq || []).push(arguments); };\n\t\tvar script = document.createElement('script');\n\t\tscript.defer = true;\n\t\tscript.src = '/_vercel/insights/script.js';\n\t\tvar head = document.querySelector('head');\n\t\thead.appendChild(script);\n\t"}],"styles":[{"type":"inline","content":"body{margin:0;font-family:Segoe UI,Tahoma,Geneva,Verdana,sans-serif;background:linear-gradient(to bottom right,#111827,#581c87,#1e3a8a);color:#fff;min-height:100vh;overflow-x:hidden}.container[data-astro-cid-qraosrxq]{max-width:600px;margin:6rem auto;padding:2rem;background:#ffffff0d;backdrop-filter:blur(16px);border:1px solid rgba(255,255,255,.1);border-radius:1.5rem;box-shadow:0 0 30px #6366f133}h1[data-astro-cid-qraosrxq]{font-size:2rem;text-align:center;margin-bottom:1.5rem;background:linear-gradient(135deg,#8b5cf6,#6366f1);-webkit-background-clip:text;-webkit-text-fill-color:transparent}form[data-astro-cid-qraosrxq]{display:flex;flex-direction:column;gap:1rem}input[data-astro-cid-qraosrxq],select[data-astro-cid-qraosrxq]{padding:.75rem 1rem;border-radius:.75rem;border:none;background:#ffffff1a;color:#fff;font-size:1rem;border:1px solid rgba(255,255,255,.2)}input[data-astro-cid-qraosrxq]::placeholder{color:#ffffff80}input[data-astro-cid-qraosrxq]:focus,select[data-astro-cid-qraosrxq]:focus{outline:none;border-color:#8b5cf6;box-shadow:0 0 0 2px #8b5cf680}button[data-astro-cid-qraosrxq]{padding:.75rem 1rem;font-weight:700;border:none;border-radius:.75rem;background:linear-gradient(135deg,#6366f1,#8b5cf6);color:#fff;cursor:pointer;transition:transform .3s ease,box-shadow .3s ease}button[data-astro-cid-qraosrxq]:hover{transform:translateY(-2px);box-shadow:0 8px 24px #6366f166}button[data-astro-cid-qraosrxq]:disabled{opacity:.6;cursor:not-allowed;transform:none}.note[data-astro-cid-qraosrxq]{text-align:center;font-size:.9rem;color:#94a3b8;margin-top:1rem}.note[data-astro-cid-qraosrxq] a[data-astro-cid-qraosrxq]{color:#8b5cf6;text-decoration:underline}.message[data-astro-cid-qraosrxq]{padding:.75rem;border-radius:.75rem;margin-bottom:1rem;text-align:center;font-weight:700}.success[data-astro-cid-qraosrxq]{background:#22c55e33;border:1px solid rgba(34,197,94,.3);color:#4ade80}.error[data-astro-cid-qraosrxq]{background:#ef444433;border:1px solid rgba(239,68,68,.3);color:#f87171}select[data-astro-cid-qraosrxq][multiple]{height:auto;min-height:100px}select[data-astro-cid-qraosrxq] option[data-astro-cid-qraosrxq]{background:#1e293b;color:#fff;padding:.5rem}label[data-astro-cid-qraosrxq]{color:#cbd5e1;font-size:.9rem;margin-bottom:.5rem}.form-group[data-astro-cid-qraosrxq]{display:flex;flex-direction:column;gap:.5rem}\n"}],"routeData":{"route":"/register","isIndex":false,"type":"page","pattern":"^\\/register\\/?$","segments":[[{"content":"register","dynamic":false,"spread":false}]],"params":[],"component":"src/pages/register.astro","pathname":"/register","prerender":false,"fallbackRoutes":[],"distURL":[],"origin":"project","_meta":{"trailingSlash":"ignore"}}},{"file":"","links":[],"scripts":[{"stage":"head-inline","children":"window.va = window.va || function () { (window.vaq = window.vaq || []).push(arguments); };\n\t\tvar script = document.createElement('script');\n\t\tscript.defer = true;\n\t\tscript.src = '/_vercel/insights/script.js';\n\t\tvar head = document.querySelector('head');\n\t\thead.appendChild(script);\n\t"}],"styles":[{"type":"external","src":"/_astro/about.yPWawPvy.css"},{"type":"inline","content":".footer-container[data-astro-cid-tnm6jnzf]{background:linear-gradient(135deg,#0a0a0a,#1a1a1a,#0f2027);color:#fff;position:relative;overflow:hidden}.footer-container[data-astro-cid-tnm6jnzf]:before{content:\"\";position:absolute;top:0;left:0;right:0;height:1px;background:linear-gradient(90deg,transparent,rgba(0,212,170,.3),transparent)}.footer-content[data-astro-cid-tnm6jnzf]{max-width:1200px;margin:0 auto;padding:4rem 2rem 2rem}.footer-grid[data-astro-cid-tnm6jnzf]{display:grid;grid-template-columns:2fr 1fr 1fr 1fr;gap:3rem;margin-bottom:3rem}.footer-section[data-astro-cid-tnm6jnzf]{animation:fadeInUp .6s ease-out}.footer-section[data-astro-cid-tnm6jnzf]:nth-child(2){animation-delay:.1s}.footer-section[data-astro-cid-tnm6jnzf]:nth-child(3){animation-delay:.2s}.footer-section[data-astro-cid-tnm6jnzf]:nth-child(4){animation-delay:.3s}.footer-logo[data-astro-cid-tnm6jnzf]{display:flex;align-items:center;gap:.75rem;margin-bottom:1.5rem}.logo-container[data-astro-cid-tnm6jnzf]{position:relative;width:48px;height:48px;flex-shrink:0}.logo-icon[data-astro-cid-tnm6jnzf]{width:100%;height:100%;filter:drop-shadow(0 0 8px rgba(99,102,241,.3))}.brand-content[data-astro-cid-tnm6jnzf]{display:flex;flex-direction:column}.brand-text[data-astro-cid-tnm6jnzf]{font-size:1.5rem;font-weight:700;color:#fff;letter-spacing:-.025em;line-height:1}.brand-tagline[data-astro-cid-tnm6jnzf]{font-size:.75rem;color:#fff9;margin-top:2px;font-weight:400}.footer-description[data-astro-cid-tnm6jnzf]{color:#ffffffb3;line-height:1.6;margin-bottom:1.5rem;font-size:.95rem}.social-links[data-astro-cid-tnm6jnzf]{display:flex;gap:1rem}.social-link[data-astro-cid-tnm6jnzf]{width:40px;height:40px;border-radius:50%;background:#ffffff0d;border:1px solid rgba(255,255,255,.1);display:flex;align-items:center;justify-content:center;color:#fff9;text-decoration:none;transition:all .3s ease;backdrop-filter:blur(10px)}.social-link[data-astro-cid-tnm6jnzf]:hover{background:#00d4aa1a;border-color:#00d4aa;color:#00d4aa;transform:translateY(-2px)}.footer-title[data-astro-cid-tnm6jnzf]{font-size:1rem;font-weight:600;margin-bottom:1rem;color:#ffffffe6}.footer-links[data-astro-cid-tnm6jnzf]{list-style:none;padding:0;margin:0}.footer-links[data-astro-cid-tnm6jnzf] li[data-astro-cid-tnm6jnzf]{margin-bottom:.75rem}.footer-links[data-astro-cid-tnm6jnzf] a[data-astro-cid-tnm6jnzf]{color:#fff9;text-decoration:none;font-size:.9rem;transition:color .3s ease}.footer-links[data-astro-cid-tnm6jnzf] a[data-astro-cid-tnm6jnzf]:hover{color:#00d4aa}.footer-bottom[data-astro-cid-tnm6jnzf]{border-top:1px solid rgba(255,255,255,.1);padding-top:2rem}.footer-bottom-content[data-astro-cid-tnm6jnzf]{display:flex;justify-content:space-between;align-items:center;flex-wrap:wrap;gap:1rem}.copyright[data-astro-cid-tnm6jnzf]{color:#ffffff80;font-size:.85rem;margin:0}.footer-legal[data-astro-cid-tnm6jnzf]{display:flex;gap:1.5rem}.footer-legal[data-astro-cid-tnm6jnzf] a[data-astro-cid-tnm6jnzf]{color:#ffffff80;text-decoration:none;font-size:.85rem;transition:color .3s ease}.footer-legal[data-astro-cid-tnm6jnzf] a[data-astro-cid-tnm6jnzf]:hover{color:#00d4aa}@media (max-width: 768px){.footer-content[data-astro-cid-tnm6jnzf]{padding:3rem 1rem 1rem}.footer-grid[data-astro-cid-tnm6jnzf]{grid-template-columns:1fr;gap:2rem;margin-bottom:2rem}.footer-section[data-astro-cid-tnm6jnzf]{text-align:center}.footer-logo[data-astro-cid-tnm6jnzf],.social-links[data-astro-cid-tnm6jnzf]{justify-content:center}.footer-bottom-content[data-astro-cid-tnm6jnzf]{flex-direction:column;text-align:center}.footer-legal[data-astro-cid-tnm6jnzf]{justify-content:center}}@media (max-width: 480px){.footer-legal[data-astro-cid-tnm6jnzf]{flex-direction:column;gap:.5rem}.logo-container[data-astro-cid-tnm6jnzf]{width:40px;height:40px}.brand-text[data-astro-cid-tnm6jnzf]{font-size:1.25rem}}@keyframes fadeInUp{0%{opacity:0;transform:translateY(20px)}to{opacity:1;transform:translateY(0)}}\n"},{"type":"external","src":"/_astro/index.BHchBIlG.css"}],"routeData":{"route":"/","isIndex":true,"type":"page","pattern":"^\\/$","segments":[],"params":[],"component":"src/pages/index.astro","pathname":"/","prerender":false,"fallbackRoutes":[],"distURL":[],"origin":"project","_meta":{"trailingSlash":"ignore"}}}],"base":"/","trailingSlash":"ignore","compressHTML":true,"componentMetadata":[["D:/project/web/tukangcs/absent-accretion/src/pages/about.astro",{"propagation":"none","containsHead":true}],["D:/project/web/tukangcs/absent-accretion/src/pages/contact.astro",{"propagation":"none","containsHead":true}],["D:/project/web/tukangcs/absent-accretion/src/pages/dashboard.astro",{"propagation":"none","containsHead":true}],["D:/project/web/tukangcs/absent-accretion/src/pages/product.astro",{"propagation":"none","containsHead":true}],["D:/project/web/tukangcs/absent-accretion/src/pages/register.astro",{"propagation":"none","containsHead":true}],["D:/project/web/tukangcs/absent-accretion/src/pages/get-started.astro",{"propagation":"none","containsHead":true}],["D:/project/web/tukangcs/absent-accretion/src/pages/login.astro",{"propagation":"none","containsHead":true}],["D:/project/web/tukangcs/absent-accretion/src/pages/index.astro",{"propagation":"none","containsHead":true}]],"renderers":[],"clientDirectives":[["idle","(()=>{var l=(n,t)=>{let i=async()=>{await(await n())()},e=typeof t.value==\"object\"?t.value:void 0,s={timeout:e==null?void 0:e.timeout};\"requestIdleCallback\"in window?window.requestIdleCallback(i,s):setTimeout(i,s.timeout||200)};(self.Astro||(self.Astro={})).idle=l;window.dispatchEvent(new Event(\"astro:idle\"));})();"],["load","(()=>{var e=async t=>{await(await t())()};(self.Astro||(self.Astro={})).load=e;window.dispatchEvent(new Event(\"astro:load\"));})();"],["media","(()=>{var n=(a,t)=>{let i=async()=>{await(await a())()};if(t.value){let e=matchMedia(t.value);e.matches?i():e.addEventListener(\"change\",i,{once:!0})}};(self.Astro||(self.Astro={})).media=n;window.dispatchEvent(new Event(\"astro:media\"));})();"],["only","(()=>{var e=async t=>{await(await t())()};(self.Astro||(self.Astro={})).only=e;window.dispatchEvent(new Event(\"astro:only\"));})();"],["visible","(()=>{var a=(s,i,o)=>{let r=async()=>{await(await s())()},t=typeof i.value==\"object\"?i.value:void 0,c={rootMargin:t==null?void 0:t.rootMargin},n=new IntersectionObserver(e=>{for(let l of e)if(l.isIntersecting){n.disconnect(),r();break}},c);for(let e of o.children)n.observe(e)};(self.Astro||(self.Astro={})).visible=a;window.dispatchEvent(new Event(\"astro:visible\"));})();"]],"entryModules":{"\u0000@astrojs-ssr-adapter":"_@astrojs-ssr-adapter.mjs","\u0000noop-middleware":"_noop-middleware.mjs","\u0000noop-actions":"_noop-actions.mjs","\u0000@astro-page:src/pages/about@_@astro":"pages/about.astro.mjs","\u0000@astro-page:src/pages/api/auth/login@_@ts":"pages/api/auth/login.astro.mjs","\u0000@astro-page:src/pages/api/auth/register@_@ts":"pages/api/auth/register.astro.mjs","\u0000@astro-page:src/pages/api/auth/signout@_@ts":"pages/api/auth/signout.astro.mjs","\u0000@astro-page:src/pages/api/test-env@_@ts":"pages/api/test-env.astro.mjs","\u0000@astro-page:src/pages/contact@_@astro":"pages/contact.astro.mjs","\u0000@astro-page:src/pages/dashboard@_@astro":"pages/dashboard.astro.mjs","\u0000@astro-page:src/pages/product@_@astro":"pages/product.astro.mjs","\u0000@astro-page:src/pages/register@_@astro":"pages/register.astro.mjs","\u0000@astrojs-ssr-virtual-entry":"entry.mjs","\u0000@astro-page:src/pages/get-started@_@astro":"pages/get-started.astro.mjs","\u0000@astro-page:src/pages/login@_@astro":"pages/login.astro.mjs","\u0000@astro-page:src/pages/index@_@astro":"pages/index.astro.mjs","\u0000@astro-renderers":"renderers.mjs","\u0000@astro-page:node_modules/astro/dist/assets/endpoint/generic@_@js":"pages/_image.astro.mjs","D:/project/web/tukangcs/absent-accretion/node_modules/astro/dist/assets/services/sharp.js":"chunks/sharp_D3uemgoL.mjs","\u0000@astrojs-manifest":"manifest_BOaE-K_2.mjs","D:/project/web/tukangcs/absent-accretion/src/pages/dashboard.astro?astro&type=script&index=0&lang.ts":"_astro/dashboard.astro_astro_type_script_index_0_lang.l0sNRNKZ.js","D:/project/web/tukangcs/absent-accretion/src/pages/about.astro?astro&type=script&index=0&lang.ts":"_astro/about.astro_astro_type_script_index_0_lang.DJCsEHvF.js","D:/project/web/tukangcs/absent-accretion/src/pages/contact.astro?astro&type=script&index=0&lang.ts":"_astro/contact.astro_astro_type_script_index_0_lang.DQM8NYso.js","D:/project/web/tukangcs/absent-accretion/src/pages/product.astro?astro&type=script&index=0&lang.ts":"_astro/product.astro_astro_type_script_index_0_lang.CFw735UV.js","D:/project/web/tukangcs/absent-accretion/src/pages/register.astro?astro&type=script&index=0&lang.ts":"_astro/register.astro_astro_type_script_index_0_lang.BOpe17A0.js","D:/project/web/tukangcs/absent-accretion/src/components/navbar/navbar.astro?astro&type=script&index=0&lang.ts":"_astro/navbar.astro_astro_type_script_index_0_lang.1ySohD_6.js","D:/project/web/tukangcs/absent-accretion/src/components/hero/HerosectionTwo.astro?astro&type=script&index=0&lang.ts":"_astro/HerosectionTwo.astro_astro_type_script_index_0_lang.kmSO1nxx.js","D:/project/web/tukangcs/absent-accretion/node_modules/@vercel/analytics/dist/astro/index.astro?astro&type=script&index=0&lang.ts":"_astro/index.astro_astro_type_script_index_0_lang.3u430bf-.js","@astrojs/svelte/client.js":"_astro/client.svelte.BpuXKReg.js","astro:scripts/before-hydration.js":""},"inlinedScripts":[["D:/project/web/tukangcs/absent-accretion/src/pages/dashboard.astro?astro&type=script&index=0&lang.ts",""],["D:/project/web/tukangcs/absent-accretion/src/pages/about.astro?astro&type=script&index=0&lang.ts","const i={threshold:.1,rootMargin:\"0px 0px -50px 0px\"},c=new IntersectionObserver(e=>{e.forEach(t=>{t.isIntersecting&&t.target.classList.add(\"visible\")})},i);document.addEventListener(\"DOMContentLoaded\",()=>{document.querySelectorAll(\".reveal-element, .reveal-left, .reveal-right, .fade-scale\").forEach(t=>c.observe(t))});document.querySelectorAll(\"button\").forEach(e=>{e.addEventListener(\"click\",t=>{const r=document.createElement(\"span\"),n=e.getBoundingClientRect(),a=Math.max(n.width,n.height),o=t.clientX-n.left-a/2,l=t.clientY-n.top-a/2;r.style.cssText=`\n                    position: absolute;\n                    width: ${a}px;\n                    height: ${a}px;\n                    left: ${o}px;\n                    top: ${l}px;\n                    background: rgba(255, 255, 255, 0.3);\n                    border-radius: 50%;\n                    transform: scale(0);\n                    animation: ripple 0.6s linear;\n                    pointer-events: none;\n                `,e.style.position=\"relative\",e.style.overflow=\"hidden\",e.appendChild(r),setTimeout(()=>r.remove(),600),alert(\"🎉 Terima kasih atas minat Anda! Tim kami akan segera menghubungi Anda.\")})});window.addEventListener(\"scroll\",()=>{const e=window.pageYOffset;document.querySelectorAll(\".bg-decoration\").forEach((r,n)=>{const a=.5+n*.1;r.style.transform=`translateY(${e*a}px)`})});const s=document.createElement(\"style\");s.innerHTML=`\n            @keyframes ripple {\n                to {\n                    transform: scale(4);\n                    opacity: 0;\n                }\n            }\n        `;document.head.appendChild(s);"],["D:/project/web/tukangcs/absent-accretion/src/pages/contact.astro?astro&type=script&index=0&lang.ts","const n=document.getElementById(\"contactForm\");n&&n.addEventListener(\"submit\",function(t){t.preventDefault();const r=new FormData(t.target);Object.fromEntries(r);const e=document.querySelector(\".submit-btn\");if(e){const o=e.textContent||\"\";e.textContent=\"Sending...\",e.disabled=!0,setTimeout(()=>{alert(\"Thank you for your message! We'll get back to you soon.\"),t.target.reset(),e.textContent=o,e.disabled=!1},1500)}});document.querySelectorAll(\".nav-arrow\").forEach(t=>{t.addEventListener(\"click\",function(){this.style.transform=\"scale(0.9)\",setTimeout(()=>{this.style.transform=\"scale(1)\"},150)})});document.querySelectorAll(\"input, textarea\").forEach(t=>{t.addEventListener(\"focus\",function(){this.parentElement&&(this.parentElement.style.transform=\"translateY(-2px)\")}),t.addEventListener(\"blur\",function(){this.parentElement&&(this.parentElement.style.transform=\"translateY(0)\")})});"],["D:/project/web/tukangcs/absent-accretion/src/pages/product.astro?astro&type=script&index=0&lang.ts","const o={threshold:.1,rootMargin:\"0px 0px -50px 0px\"},c=new IntersectionObserver(t=>{t.forEach(e=>{e.isIntersecting&&e.target.classList.add(\"visible\")})},o);document.addEventListener(\"DOMContentLoaded\",()=>{document.querySelectorAll(\".reveal-element, .reveal-left, .reveal-right\").forEach(e=>c.observe(e))});document.querySelectorAll(\"button\").forEach(t=>{t.addEventListener(\"click\",e=>{const s=document.createElement(\"span\"),a=t.getBoundingClientRect(),n=Math.max(a.width,a.height),r=e.clientX-a.left-n/2,i=e.clientY-a.top-n/2;s.style.cssText=`\n                    position: absolute;\n                    width: ${n}px;\n                    height: ${n}px;\n                    left: ${r}px;\n                    top: ${i}px;\n                    background: rgba(255, 255, 255, 0.3);\n                    border-radius: 50%;\n                    transform: scale(0);\n                    animation: ripple 0.6s linear;\n                    pointer-events: none;\n                `,t.style.position=\"relative\",t.style.overflow=\"hidden\",t.appendChild(s),setTimeout(()=>s.remove(),600);const l=t.textContent||\"\";l.includes(\"Gratis\")?alert(\"🎉 Selamat! Anda akan diarahkan ke halaman pendaftaran trial gratis.\"):l.includes(\"Konsultasi\")||l.includes(\"Sales\")?alert(\"📞 Tim sales kami akan menghubungi Anda dalam 1x24 jam untuk konsultasi gratis.\"):l.includes(\"Mulai\")&&alert(\"🚀 Terima kasih! Silakan pilih paket yang sesuai dengan kebutuhan bisnis Anda.\")})});window.addEventListener(\"scroll\",()=>{const t=window.pageYOffset;document.querySelectorAll(\".bg-decoration\").forEach((s,a)=>{const n=.5+a*.1;s.style.transform=`translateY(${t*n}px)`})});document.querySelectorAll(\".product-card\").forEach(t=>{const e=t;e.addEventListener(\"mouseenter\",()=>{e.style.transform=e.style.transform===\"scale(1.05)\"?\"translateY(-5px) scale(1.05)\":\"translateY(-5px) scale(1.02)\"}),e.addEventListener(\"mouseleave\",()=>{e.style.transform=e.style.transform.includes(\"scale(1.05)\")?\"scale(1.05)\":\"scale(1)\"})});"],["D:/project/web/tukangcs/absent-accretion/src/pages/register.astro?astro&type=script&index=0&lang.ts","function a(r,s){const e=document.getElementById(\"message\");e&&(e.innerHTML=`<div class=\"message ${s}\">${r}</div>`,s===\"success\"&&setTimeout(()=>{e.innerHTML=\"\"},3e3))}function d(r){const s=r.get(\"email\"),e=r.get(\"password\"),o=r.get(\"nama_umkm\"),n=r.get(\"sektor\");return!s||!e||!o||!n?(a(\"Semua field wajib diisi!\",\"error\"),!1):/^[^\\s@]+@[^\\s@]+\\.[^\\s@]+$/.test(s)?e.length<6?(a(\"Password minimal 6 karakter!\",\"error\"),!1):!0:(a(\"Format email tidak valid!\",\"error\"),!1)}const l=document.getElementById(\"register-form\");l?.addEventListener(\"submit\",async r=>{r.preventDefault();const s=document.getElementById(\"submit-btn\"),e=new FormData(r.target);if(!d(e))return;const o=document.getElementById(\"sosmed\"),n=o?Array.from(o.selectedOptions).map(t=>t.value):[];e.delete(\"sosmed\"),n.forEach(t=>{e.append(\"sosmed\",t)}),s&&(s.disabled=!0,s.textContent=\"Mendaftar...\");try{let t=await fetch(\"/api/auth/register\",{method:\"POST\",body:e});if(!t.ok&&t.status===400){const m={nama_umkm:e.get(\"nama_umkm\"),email:e.get(\"email\"),password:e.get(\"password\"),sektor:e.get(\"sektor\"),sosmed:n};t=await fetch(\"/api/auth/register\",{method:\"POST\",headers:{\"Content-Type\":\"application/json\"},body:JSON.stringify(m)})}const i=await t.json();t.ok?(a(\"Registrasi berhasil! Silakan cek email untuk verifikasi akun Anda.\",\"success\"),r.target?.reset(),setTimeout(()=>{window.location.href=\"/login\"},3e3)):a(`Error: ${i.error}`,\"error\")}catch(t){console.error(\"Registration error:\",t),a(`Network error: ${t.message}`,\"error\")}finally{s&&(s.disabled=!1,s.textContent=\"Daftar Sekarang\")}});"],["D:/project/web/tukangcs/absent-accretion/src/components/navbar/navbar.astro?astro&type=script&index=0&lang.ts","let s=!1,c=!1;const m=()=>{c||(requestAnimationFrame(()=>{const e=document.getElementById(\"navbar\"),t=window.scrollY>20;t!==s&&(s=t,e&&e.classList.toggle(\"scrolled\",s)),c=!1}),c=!0)};let n=!1;const a=()=>{n=!n;const e=document.getElementById(\"mobile-nav\"),t=document.getElementById(\"mobile-menu-btn\");e&&t&&(e.classList.toggle(\"open\",n),t.classList.toggle(\"active\",n),document.body.style.overflow=n?\"hidden\":\"\")},l=()=>{if(n){n=!1;const e=document.getElementById(\"mobile-nav\"),t=document.getElementById(\"mobile-menu-btn\");e&&t&&(e.classList.remove(\"open\"),t.classList.remove(\"active\"),document.body.style.overflow=\"\")}};document.addEventListener(\"DOMContentLoaded\",()=>{window.addEventListener(\"scroll\",m,{passive:!0});const e=document.getElementById(\"mobile-menu-btn\");e&&e.addEventListener(\"click\",a),document.querySelectorAll(\".mobile-link\").forEach(o=>{o.addEventListener(\"click\",l)});const d=document.querySelector(\".mobile-cta\");d&&d.addEventListener(\"click\",l);const i=document.getElementById(\"mobile-nav\");i&&i.addEventListener(\"click\",o=>{o.target&&(o.target===i||o.target.classList.contains(\"mobile-bg-blur\"))&&l()}),document.addEventListener(\"keydown\",o=>{o.key===\"Escape\"&&n&&l()})});window.addEventListener(\"beforeunload\",()=>{window.removeEventListener(\"scroll\",m),document.body.style.overflow=\"\"});"],["D:/project/web/tukangcs/absent-accretion/src/components/hero/HerosectionTwo.astro?astro&type=script&index=0&lang.ts","const s={threshold:.1,rootMargin:\"0px 0px -50px 0px\"},o=new IntersectionObserver(e=>{e.forEach(t=>{t.isIntersecting&&t.target.classList.add(\"visible\")})},s);document.addEventListener(\"DOMContentLoaded\",()=>{document.querySelectorAll(\".reveal-element, .reveal-left, .reveal-right\").forEach(t=>o.observe(t))});document.querySelectorAll(\"button\").forEach(e=>{e.addEventListener(\"click\",t=>{const l=document.createElement(\"span\"),a=e.getBoundingClientRect(),n=Math.max(a.width,a.height),i=t.clientX-a.left-n/2,r=t.clientY-a.top-n/2;l.style.cssText=`\n                    position: absolute;\n                    width: ${n}px;\n                    height: ${n}px;\n                    left: ${i}px;\n                    top: ${r}px;\n                    background: rgba(255, 255, 255, 0.3);\n                    border-radius: 50%;\n                    transform: scale(0);\n                    animation: ripple 0.6s linear;\n                    pointer-events: none;\n                `,e.style.position=\"relative\",e.style.overflow=\"hidden\",e.appendChild(l),setTimeout(()=>l.remove(),600),e.textContent&&e.textContent.includes(\"Demo\")?alert(\"🚀 Demo akan segera dimulai! Fitur ini akan menampilkan cara kerja TukangCS.\"):e.textContent&&e.textContent.includes(\"Konsultasi\")?alert(\"📅 Terima kasih! Tim kami akan menghubungi Anda untuk jadwal konsultasi gratis.\"):e.textContent&&e.textContent.includes(\"Mulai\")?alert(\"🎉 Selamat datang di TukangCS! Mari mulai perjalanan transformasi digital Anda.\"):e.textContent&&e.textContent.includes(\"Pelajari\")&&alert(\"📚 Dokumentasi lengkap akan segera tersedia untuk Anda pelajari.\")})});window.addEventListener(\"scroll\",()=>{const e=window.pageYOffset;document.querySelectorAll(\".bg-decoration\").forEach((l,a)=>{const n=.5+a*.1;l.style.transform=`translateY(${e*n}px)`})});"],["D:/project/web/tukangcs/absent-accretion/node_modules/@vercel/analytics/dist/astro/index.astro?astro&type=script&index=0&lang.ts","var f=\"@vercel/analytics\",l=\"1.5.0\",w=()=>{window.va||(window.va=function(...r){(window.vaq=window.vaq||[]).push(r)})};function d(){return typeof window<\"u\"}function u(){try{const e=\"production\"}catch{}return\"production\"}function v(e=\"auto\"){if(e===\"auto\"){window.vam=u();return}window.vam=e}function m(){return(d()?window.vam:u())||\"production\"}function c(){return m()===\"development\"}function b(e,r){if(!e||!r)return e;let n=e;try{const t=Object.entries(r);for(const[a,i]of t)if(!Array.isArray(i)){const o=s(i);o.test(n)&&(n=n.replace(o,`/[${a}]`))}for(const[a,i]of t)if(Array.isArray(i)){const o=s(i.join(\"/\"));o.test(n)&&(n=n.replace(o,`/[...${a}]`))}return n}catch{return e}}function s(e){return new RegExp(`/${h(e)}(?=[/?#]|$)`)}function h(e){return e.replace(/[.*+?^${}()|[\\]\\\\]/g,\"\\\\$&\")}function y(e){return e.scriptSrc?e.scriptSrc:c()?\"https://va.vercel-scripts.com/v1/script.debug.js\":e.basePath?`${e.basePath}/insights/script.js`:\"/_vercel/insights/script.js\"}function g(e={debug:!0}){var r;if(!d())return;v(e.mode),w(),e.beforeSend&&((r=window.va)==null||r.call(window,\"beforeSend\",e.beforeSend));const n=y(e);if(document.head.querySelector(`script[src*=\"${n}\"]`))return;const t=document.createElement(\"script\");t.src=n,t.defer=!0,t.dataset.sdkn=f+(e.framework?`/${e.framework}`:\"\"),t.dataset.sdkv=l,e.disableAutoTrack&&(t.dataset.disableAutoTrack=\"1\"),e.endpoint?t.dataset.endpoint=e.endpoint:e.basePath&&(t.dataset.endpoint=`${e.basePath}/insights`),e.dsn&&(t.dataset.dsn=e.dsn),t.onerror=()=>{const a=c()?\"Please check if any ad blockers are enabled and try again.\":\"Be sure to enable Web Analytics for your project and deploy again. See https://vercel.com/docs/analytics/quickstart for more information.\";console.log(`[Vercel Web Analytics] Failed to load script from ${n}. ${a}`)},c()&&e.debug===!1&&(t.dataset.debug=\"false\"),document.head.appendChild(t)}function p({route:e,path:r}){var n;(n=window.va)==null||n.call(window,\"pageview\",{route:e,path:r})}function k(){try{return}catch{}}customElements.define(\"vercel-analytics\",class extends HTMLElement{constructor(){super();try{const r=JSON.parse(this.dataset.props??\"{}\"),n=JSON.parse(this.dataset.params??\"{}\");g({...r,disableAutoTrack:!0,framework:\"astro\",basePath:k(),beforeSend:window.webAnalyticsBeforeSend});const t=this.dataset.pathname;p({route:b(t??\"\",n),path:t})}catch(r){throw new Error(`Failed to parse WebAnalytics properties: ${r}`)}}});"]],"assets":["/_astro/about.yPWawPvy.css","/_astro/about.CNm0yoeI.css","/_astro/contact.BeCkmT8s.css","/_astro/index.BHchBIlG.css","/_astro/product.DkOdxu3m.css","/favicon.svg","/images/app-preview.svg","/images/app-store.svg","/images/difa-founder.png","/images/google-play.svg","/images/hero-teks-onframe.png","/images/hero-teks-onframe.svg","/images/isaac-co-founder.png","/images/main-char.png","/images/main-char.svg","/images/platform.png","/images/socialmedia.svg","/images/working-girl.png","/images/working-girl.svg","/_astro/client.svelte.BpuXKReg.js"],"buildFormat":"directory","checkOrigin":true,"serverIslandNameMap":[],"key":"uxM4t9bwKPfHVVeBEgwLo+XQGjF80KcwF5cVkFteJSw="});
-if (manifest.sessionConfig) manifest.sessionConfig.driverModule = null;
-
-export { manifest };
