+++ .vercel/output/_functions/pages/api/auth/register.astro.mjs
@@ -1,4 +1,4 @@
-import { a as adminAuth, b as adminDb } from '../../../chunks/firebase-admin_kE6dM-1a.mjs';
+import { a as adminAuth, b as adminDb } from '../../../chunks/firebase-admin_CZtLfzeI.mjs';
 export { renderers } from '../../../renderers.mjs';
 
 const prerender = false;
